dodanie pobierania hasla bazy z get_secret, zaktualizowanie istniejacych testow i ich struktury katalogow, wszystko dziala
